Added mapper 033 - doesn't work perfectly

+++ src/mapper.cpp
@@ -1,7 +1,7 @@
 
 #include "mapper.h"
 
-Mapper::Mapper(uint8_t p, uint8_t c)
+Mapper::Mapper(uint16_t p, uint16_t c)
 {
     prgBanks = p;
     chrBanks = c;
